Collected types in `pagedid.d.ts`
